requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
easyocr>=1.7.0          # OCR cho biển số (nhẹ, hỗ trợ tiếng Việt)
python-dotenv>=0.19.0
roboflow>=1.1.0